+++ src/translations/zh/elementmessages.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * Element Messages plugin for CraftCMS 3
+ *
+ * @link      https://panlatent.com/
+ * @copyright Copyright (c) 2019 panlatent@gmail.com
+ */
+
+return [
+    'Message' => '消息',
+    'Messages' => '消息',
+    'Settings' => '设置',
+    'No message exists yet.' => '尚不存在消息。',
+    'Name' => '名称',
+    'Sender' => '发送者',
+    'Target' => '目标',
+    'Content' => '内容',
+    'Post Date' => '发布日期',
+    '{sender} sends {content} to {target}' => '{sender} 将 {content} 发送到 {target}',
+    '{sender} sends empty content to {target}' => '{sender} 将空内容发送到 {target}',
+    'Are you sure you want to delete this message?' => '你确定要删除这个消息？',
+    'General Settings' => '常规设置',
+    'CP Nav Name' => 'CP 导航栏名称',
+    'Messages Page Size' => '每页消息数量（消息页）',
+    'Couldn’t save settings.' => '无法保存设置。',
+    'Settings saved.' => '设置已保存。',
+];