+++ src/user/Permissions.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * Element Messages plugin for CraftCMS 3
+ *
+ * @link      https://panlatent.com/
+ * @copyright Copyright (c) 2019 panlatent@gmail.com
+ */
+
+namespace panlatent\elementmessages\user;
+
+/**
+ * Class Permissions
+ *
+ * @package panlatent\elementmessages\user
+ * @author Panlatent <panlatent@gmail.com>
+ */
+abstract class Permissions
+{
+    // Constants
+    // =========================================================================
+
+    const MANAGE_MESSAGES = 'elementmessages-manageMessages';
+    const MANAGE_SETTINGS = 'elementmessages-manageSettings';
+}