+++ CHANGELOG.md
@@ -11,6 +11,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Updated
 ### Fixed
 
+## 0.1.1 - 2019-07-02
+
+### Added
+- Add Cp messages view.
+- Add Cp plugin settings.
+
+### Fixed
+- Fixed message content object not allow null.
+
 ## 0.1.0 - 2019-04-15
 
 Init the plugin.